% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMediansAll.R
\name{calcMediansAll}
\alias{calcMediansAll}
\title{Calculate cluster medians across all samples (for plotting)}
\usage{
calcMediansAll(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = markers, assay = values (marker expression values).
}
\description{
Calculate cluster medians (median expression of each marker) for each cluster
}
\details{
Calculate median expression of each marker for each cluster across all samples.

The data object is assumed to contain a vector \code{is_marker_col} in the column
meta-data (see \code{\link{prepareData}}), which indicates whether each column
represents a marker. Cluster medians are calculated for these columns.

The cluster medians (across all samples) are required for plotting functions.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = markers, assay = values (marker expression
values).
}
\examples{
# See full examples in testing functions.

}
\seealso{
to do
}
