% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMediansAndFreq.R
\name{calcMediansAndFreq}
\alias{calcMediansAndFreq}
\title{Calculate cluster medians and frequencies}
\usage{
calcMediansAndFreq(d_se)
}
\arguments{
\item{d_se}{Transformed data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, with cluster labels 
added in row meta-data using \code{\link{generateClusters}}.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = functional markers. The 
  additional last sheet ('assay' slot) contains the cluster frequencies.
}
\description{
Calculate cluster medians (median functional marker expression) and frequencies (number
of cells) by cluster and sample
}
\details{
Calculate median expression of functional markers (cluster medians) and number of cells 
(cluster frequencies) by cluster and sample (i.e. for each cluster in each sample).

The cluster frequencies are used as weights in the subsequent statistical tests. The
cluster medians are either used for directly testing differences in medians
('diffcyt-med'), or for visualization of results from the other methodologies
('diffcyt-FDA' and 'diffcyt-KS').

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} 
object, where rows = clusters, columns = samples, sheets ('assay' slots) = functional 
markers. The additional last sheet ('assay' slot) contains the cluster frequencies.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
