% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSubsetVals.R
\name{calcSubsetVals}
\alias{calcSubsetVals}
\title{Subset values for each cluster-sample combination}
\usage{
calcSubsetVals(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}). Column
meta-data is assumed to contain vectors \code{is_marker_col}, \code{is_identity_col},
and \code{is_func_col}.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, sheets ('assay' slots) = markers. Each entry is a vector
  of expression values with length equal to the number of cells. The \code{metadata}
  slot contains variables \code{id_identity_markers} and \code{id_func_markers}, which
  can be accessed with \code{metadata(d_medians)$id_identity_markers} and
  \code{metadata(d_medians)$id_func_markers}.
}
\description{
Subset values for each cluster-sample combination for each marker, required for
differential functional state testing
}
\details{
Subsets the expression values for each cluster-sample combination for each marker, and
returns as a new \code{\link{SummarizedExperiment}} object. The values can then be
provided to the differential functional state testing functions.

The data object is assumed to contain vectors \code{is_marker_col},
\code{is_identity_col}, and \code{is_func_col} in the column meta-data (see
\code{\link{prepareData}}). These indicate the sets of all marker columns, identity
marker columns, and functional marker columns. Cluster medians are calculated for all
markers.

The expression values are returned in a new
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = clusters,
columns = samples, sheets ('assay' slots) = markers. Note that there is a separate
table of values ('assay') for each marker, and each 'value' in the tables consists of a
vector of expression values. The \code{metadata} slot also contains variables
\code{id_identity_markers} and \code{id_func_markers}, which can be used to identify
the sets of identity and functional markers.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
