% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFreq.R
\name{calculateFreq}
\alias{calculateFreq}
\title{Calculate cluster frequencies}
\usage{
calculateFreq(d_transf, clus)
}
\arguments{
\item{d_transf}{Transformed input data from the previous steps. This should be in the 
form of a \code{\link[flowCore]{flowSet}} object from the 
\code{\link[flowCore]{flowCore}} package.}

\item{clus}{Cluster labels for individual cells, from \code{\link{generateClusters}}.}
}
\value{
List containing:
\itemize{
\item tbl_freq: Matrix of cluster frequencies or abundances (rows = clusters, columns 
= samples).
\item tbl_prop: Matrix of cluster proportions (rows = clusters, columns = samples).
}
}
\description{
Calculate cluster frequencies (number of cells per cluster)
}
\details{
Calculate number of cells per cluster (cluster frequencies or abundances). This is 
required for subsequent differential tests, e.g. differential abundance of clusters or
differential expression of functional markers using frequencies as weights.

This function should be run after generating clusters with 
\code{\link{generateClusters}}.
}
\examples{
# need to create a small example data set for examples
}
