% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{checkInputData}
\alias{checkInputData}
\title{Check input data format}
\usage{
checkInputData(d_input)
}
\arguments{
\item{d_input}{Input data.}
}
\value{
d_input Input data, which has been confirmed as being in the correct format 
  (\code{\link[flowCore]{flowSet}} from the \code{\link[flowCore]{flowCore}} package).
}
\description{
Check input data is in the format required by 'diffcyt' functions
}
\details{
The functions in the 'diffcyt' package assume that input data is provided as as 
\code{\link[flowCore]{flowSet}} object (a set of \code{\link[flowCore]{flowFrame}} 
object) from the \code{\link[flowCore]{flowCore}} package.

This function checks whether input data is in the correct format, and returns an error
message if not.
}
\examples{
# need to create a small example data set for examples
}
