% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_case.R
\name{complete_case}
\alias{complete_case}
\title{Complete case analysis}
\usage{
complete_case(
  data,
  censored_variable,
  censoring_indicator,
  formula,
  regression_type = c("lm", "glm", "glmer"),
  weights = "weights",
  family = "binomial"
)
}
\arguments{
\item{data}{'data.frame'}

\item{censored_variable}{name of column containing censored data}

\item{censoring_indicator}{name of column containing indication if observed
(1) or censored (0) value in column 'censored_variable'}

\item{formula}{the formula for fitting the regression model (e.g. formula(y~x))}

\item{regression_type}{function. The regression type to be used, lm for linear
regression, glm for general linear regression, glmer for generalized
linear mixed-effects models. Default: lm}

\item{weights}{name of column containing weights to be used in fitting the
regression model. Default = 'weights'. Ignored if no weights used.#'}

\item{family}{The family to be used in the regression model.
Default = "binomial". Omited if linear model is used.}
}
\value{
list of elements 'data', 'betasMean' (mean regression coef of censored
 covariate), 'betasVar' (mean variance of regression coef of censored covariate),
 'fits' (regression fit)
}
\description{
Complete case analysis
}
\examples{
 lm_formula <- formula(Y ~ Surv(X,I) + Z)
 data <- simulate_data(100, lm_formula, type = "lm", n_levels_fixeff=2)
 cc_out <- complete_case(data,"X","I",Y~X+Z)
}
