% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_multiple_imputation.R
\name{conditional_multiple_imputation}
\alias{conditional_multiple_imputation}
\title{Conditional multiple imputation}
\usage{
conditional_multiple_imputation(
  data,
  formula,
  regression_type = c("lm", "glm", "glmer"),
  repetitions = 10,
  method_est = c("mrl", "rs", "km", "cc", "pmm"),
  weights = NULL,
  contrasts = NULL,
  family = "binomial",
  id = NULL,
  verbose = FALSE,
  n_obs_min = 2,
  ...
)
}
\arguments{
\item{data}{'data.frame'}

\item{formula}{the formula for fitting the regression model with a special
syntax for the censored covariate : e.g. 'y~Surv(x,I)' means 'y~x' with 'x' being
censored and 'I' the censoring indicator (0=censored,1=observed).}

\item{regression_type}{function. The regression type to be used, lm for linear
regression, glm for general linear regression, glmer for generalized
linear mixed-effects models. Default: lm}

\item{repetitions}{number of repetitions for multiple imputation. Default: 10}

\item{method_est}{which method should be used in the imputation step. One of
'cc', 'pmm', 'mrl', 'rs', 'km'. See details.}

\item{weights}{Weights to be used in fitting the regression model. Default = NULL}

\item{contrasts}{Contrast vector to be used in testing the regression model. Default = NULL}

\item{family}{The family to be used in the regression model. Default = "binomial". Omited if linear model is used.
for the censored values.}

\item{id}{name of column containing id of sample}

\item{verbose}{Logical.}

\item{n_obs_min}{minimum number of observed events needed. default = 2.
if lower than this value will throw an error.}

\item{...}{additional arguments passed to testing}
}
\description{
Imputes censored values according to \href{https://www.researchgate.net/publication/319246304_Improved_conditional_imputation_for_linear_regression_with_a_randomly_censored_predictor}{Atem et al. 2017}
}
\details{
Possible methods in 'methods_est' are:
\describe{
  \item{'cc'}{complete case}
  \item{'pmm'}{predictive mean matching, treating censored values as missing}
  \item{'mrl'}{Mean Residual Life (Conditional single imputation from \href{https://www.researchgate.net/publication/319246304_Improved_conditional_imputation_for_linear_regression_with_a_randomly_censored_predictor}{Atem et al. 2017})}
  \item{'rs'}{Risk Set imputation}
  \item{'km'}{Kaplan Meier imputation}
}
}
\examples{
 lm_formula <- formula(Y ~ Surv(X,I) + Z)
 data <- simulate_data(100, lm_formula, type = "lm", n_levels_fixeff=2)
 cmi_out <- conditional_multiple_imputation(data,lm_formula)
}
