% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_multiple_imputation.R
\name{conditional_multiple_imputation_testing}
\alias{conditional_multiple_imputation_testing}
\title{perform fitting of imputed datasets}
\usage{
conditional_multiple_imputation_testing(
  data,
  imputed_datasets,
  censored_variable,
  censoring_indicator,
  response = NULL,
  covariates = NULL,
  id = NULL,
  formula = NULL,
  regression_type = c("lm", "glm", "glmer"),
  repetitions = 10,
  method_est = c("mrl", "rs", "km"),
  verbose = FALSE,
  weights = NULL,
  contrasts = NULL,
  family = "binomial",
  ...
)
}
\arguments{
\item{data}{'data.frame'}

\item{imputed_datasets}{all the imputed datasets for which testing is to be
performed.}

\item{censored_variable}{name of column containing censored data}

\item{censoring_indicator}{name of column containing indication if observed
(1) or censored (0) value in column 'censored_variable'}

\item{response}{name of column containing the response (can be 'NULL')}

\item{covariates}{name(s) of column(s) containing the covariate that influences
censoring}

\item{id}{name of column containing id of sample}

\item{formula}{the formula for fitting the regression model with a special
syntax for the censored covariate : e.g. 'y~Surv(x,I)' means 'y~x' with 'x' being
censored and 'I' the censoring indicator (0=censored,1=observed).}

\item{regression_type}{function. The regression type to be used, lm for linear
regression, glm for general linear regression, glmer for generalized
linear mixed-effects models. Default: lm}

\item{repetitions}{number of repetitions for multiple imputation. Default: 10}

\item{method_est}{which method should be used in the imputation step. One of
'cc', 'pmm', 'mrl', 'rs', 'km'. See details.}

\item{verbose}{Logical.}

\item{weights}{Weights to be used in fitting the regression model. Default = NULL}

\item{contrasts}{Contrast vector to be used in testing the regression model. Default = NULL}

\item{family}{The family to be used in the regression model. Default = "binomial". Omited if linear model is used.
for the censored values.}

\item{...}{additional arguments passed to testing}
}
\description{
perform fitting of imputed datasets
}
