% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_imputation.R
\name{conditional_single_imputation}
\alias{conditional_single_imputation}
\title{Conditional single imputation}
\usage{
conditional_single_imputation(
  data,
  censored_variable,
  censoring_indicator,
  response = NULL,
  covariates = NULL,
  id = NULL,
  method_est = c("mrl", "rs", "km"),
  verbose = FALSE
)
}
\arguments{
\item{data}{'data.frame'}

\item{censored_variable}{name of column containing censored data}

\item{censoring_indicator}{name of column containing indication if observed
(1) or censored (0) value in column 'censored_variable'}

\item{response}{name of column containing the response (can be 'NULL')}

\item{covariates}{name(s) of column(s) containing the covariate that influences
censoring}

\item{id}{name of column containing id of sample}

\item{method_est}{one of 'mrl','rs','km'. See \code{\link{conditional_multiple_imputation}}}

\item{verbose}{Logical}
}
\description{
Imputes censored values according to \href{https://www.researchgate.net/publication/319246304_Improved_conditional_imputation_for_linear_regression_with_a_randomly_censored_predictor}{Atem et al. 2017}
}
