% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_processing.R
\name{create_glmm_formula}
\alias{create_glmm_formula}
\title{create formula from formula containing 'Surv()'}
\usage{
create_glmm_formula(formula)
}
\arguments{
\item{formula}{formula containing indication 'Surv()'.
See \code{\link{testDA_censoredGLMM}}}
}
\value{
formula containing no 'Surv()' terms.
}
\description{
create formula from formula containing 'Surv()'
}
\examples{
  cens_formula <- formula(Y ~ Surv(X,I) + Z)
  create_glmm_formula(cens_formula)
}
