% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrast.R
\name{createContrast}
\alias{createContrast}
\title{Create contrast matrix}
\usage{
createContrast(group_IDs, contrast = NULL)
}
\arguments{
\item{group_IDs}{Vector or factor of group membership labels for each sample (e.g.
diseased vs. healthy, or treated vs. untreated). Vectors are converted to factors
internally. The first level of the factor (e.g. healthy) or first entry of the vector
will be used as the reference level for differential testing. To re-order factor
levels, use \code{\link[stats]{relevel}} or \code{\link[base]{factor}}.}

\item{contrast}{Vector defining the contrast of interest. For example, \code{c(0, 1)}
to compare the second level against the first level for a \code{group_IDs} vector of
length 2, with no other columns in the design matrix. See details for additional
explanation. Default is to compare the second level against the first level of
\code{group_IDs}, assuming no other columns in the design matrix.}
}
\value{
Returns a contrast matrix containing the contrast of interest, formatted as a
  matrix with a single column.
}
\description{
Create contrast matrix for differential testing
}
\details{
Creates a contrast matrix specifying the comparison of interest, in the correct format
for the differential testing functions. This can then be provided to the differential
testing functions, together with either a design matrix or model formula, and the data
object.

The argument \code{contrast} defines the contrast of interest, given the levels of
\code{group_IDs} and any other fixed effect terms in the design matrix or model
formula. Note that the design matrix (\code{\link{createDesignMatrix}}) or model
formula (\code{\link{createFormula}}) is constructed assuming that the first level of
\code{group_IDs} is the reference level. The design matrix is specified with an
intercept term, so that the first column represents the intercept, and the following
columns represent the difference between levels of \code{group_IDs} and the reference
level.

For example, to compare the second level against the first level for a \code{group_IDs}
vector of length 2, with no other columns in the design matrix, the \code{contrast}
argument would be \code{c(0, 1)}. To compare the second level against the first level
for a \code{group_IDs} vector of length 3, it would be \code{c(0, 1, 0)}. If there are
additional fixed effect terms in the design matrix (such as block IDs), which are not
of interest for inference, extra zeros should be included for the corresponding
additional columns of the design matrix.

The contrast matrix is formatted as a matrix with a single column containing the
contrast of interest. To perform tests for multiple contrasts, run this function and
the corresponding differential testing function multiple times.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
