% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesignMatrix.R
\name{createDesignMatrix}
\alias{createDesignMatrix}
\title{Create design matrix}
\usage{
createDesignMatrix(group_IDs, batch_IDs = NULL, covariates = NULL,
  block_IDs = NULL)
}
\arguments{
\item{group_IDs}{Vector or factor of group membership labels for each sample (e.g.
diseased vs. healthy, or treated vs. untreated). Vectors are converted to factors
internally. The first level of the factor (e.g. healthy) or first entry of the vector
will be used as the reference level for differential testing. To re-order factor
levels, use \code{\link[stats]{relevel}} or \code{\link[base]{factor}}.}

\item{batch_IDs}{(Optional) Vector or factor of batch IDs. Batch IDs are included as
columns of indicator variables in the design matrix. This allows batch effects to be
estimated during model fitting, and taken into account during inference on the group
ID parameters of interest.}

\item{covariates}{(Optional) Numeric matrix of continuous covariates. Covariates are
included as columns in the design matrix. This allows their effects to be estimated
during model fitting, and taken into account during inference on the group ID
parameters of interest.}

\item{block_IDs}{(Optional) Vector or factor of block IDs, e.g. for paired designs
(e.g. one diseased and one healthy sample per patient). If provided, block IDs are
included as fixed effects in the design matrix. Note that block IDs can also be
included as random effects by using \code{\link{createFormula}} instead; or, for some
methods, by providing them directly to the testing function (e.g.
'diffcyt-DA-limma').}
}
\value{
Returns a design matrix (numeric matrix), with one row per sample, and one
  model parameter per column.
}
\description{
Create design matrix for model fitting
}
\details{
Creates a design matrix specifying the models to be fitted. Similar to
\code{\link{createFormula}}, but returns a design matrix instead of model formula.

The design matrix can then be provided to the differential testing functions, together
with the data object and contrast matrix.

The \code{group_IDs} input specifies the groups for differential testing. This can be
provided as a vector or factor. The first level of the factor or first entry of the
vector is used as the 'reference level' for differential testing. To set a different
reference level, re-order the levels using \code{\link[stats]{relevel}} or
\code{\link[base]{factor}}.
}
\examples{
# to do

}
\seealso{
to do
}
