% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_processing.R
\name{extract_variables_from_formula}
\alias{extract_variables_from_formula}
\title{extract individuall components of 'Surv()' formula}
\usage{
extract_variables_from_formula(formula)
}
\arguments{
\item{formula}{formula containing indication 'Surv()'.
See \code{\link{testDA_censoredGLMM}}}
}
\value{
list with elements 'censored_variable', 'censoring_indicator',
 'response', 'covariates', 'random_covariates'. each a character vector.
}
\description{
extract individuall components of 'Surv()' formula
}
