% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateClusters.R
\name{generateClusters}
\alias{generateClusters}
\title{Generate clusters}
\usage{
generateClusters(d_transf, cols_to_use = NULL, xdim = 20, ydim = 20,
  seed = NULL, plot = TRUE, ...)
}
\arguments{
\item{d_transf}{Transformed input data. Data should be transformed prior to clustering
(see \code{\link{transformData}}). Input data must be in the form of a 
\code{\link[flowCore]{flowSet}} object from the \code{\link[flowCore]{flowCore}} 
package, which can be checked with \code{\link{checkInputData}}.}

\item{cols_to_use}{Columns to use for clustering. This will depend on the data set, 
but in most cases will be the set of lineage markers. Default = NULL, in which case
all columns will be used.}

\item{xdim}{Width of grid for self-organizing map for FlowSOM clustering (number of 
clusters = \code{xdim} * \code{ydim}). Default value = 20 (400 clusters).}

\item{ydim}{Height of grid for self-organizing map for FlowSOM clustering (number of 
clusters = \code{xdim} * \code{ydim}). Default value = 20 (400 clusters).}

\item{seed}{Random seed (set to an integer value for reproducible results). Default 
value = NULL.}

\item{plot}{Whether to save plot. Default = TRUE.}

\item{...}{Other parameters to pass to FlowSOM (see \code{\link[FlowSOM]{BuildSOM}} 
and \code{\link[FlowSOM]{SOM}} for details.)}
}
\value{
clus List containing cluster labels for each cell (one list item per sample).
}
\description{
Generate clusters for 'diffcyt' analysis
}
\details{
Runs a clustering algorithm to divide cells into a large number of small clusters, 
which will then be further analyzed for differential abundance or differential 
expression of functional markers.

We use the \code{\link[FlowSOM]{FlowSOM}} clustering algorithm by default (Van Gassen 
et al. 2015, \emph{Cytometry Part A}). We previously showed that \code{FlowSOM} gives
very good clustering performance for high-dimensional cytometry data, for both major
and rare cell populations, and is extremely fast (Weber and Robinson, 2016, 
\emph{Cytometry Part A}).

The clustering is run at high resolution to give a large number of small clusters, to
ensure that subtle differential expression signals in small subsets are not missed.

In most cases, only lineage markers should be used for clustering.

Note that data should be transformed prior to clustering, which can be done with the
function \code{\link{transformData}}.
}
\examples{
# need to create a small example data set for examples
}
