% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_processing.R
\name{is_valid_censored_formula}
\alias{is_valid_censored_formula}
\title{check if 'Surv()' formula is valid}
\usage{
is_valid_censored_formula(formula, throw_error = TRUE)
}
\arguments{
\item{formula}{formula containing indication 'Surv()'.
See \code{\link{testDA_censoredGLMM}}}

\item{throw_error}{should an error be thrown if 'formula' is incorrect.
default = 'TRUE'.}
}
\value{
logical
}
\description{
check if 'Surv()' formula is valid
}
\examples{
# correct formula:
cens_formula <- formula(Y ~ Surv(X,I) + Z)
is_valid_censored_formula(cens_formula)

# wrong formulas:
wrong_cens_formula_1 <- formula( ~ Surv(X,I) + Z) # no response
is_valid_censored_formula(wrong_cens_formula_1)
wrong_cens_formula_2 <- formula(Y ~ X + Z) # no censored variable
is_valid_censored_formula(wrong_cens_formula_2)
}
