% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_residual_life_imputation.R
\name{mean_residual_life_imputation}
\alias{mean_residual_life_imputation}
\title{mean residual life imputation, aka conditional imputation}
\usage{
mean_residual_life_imputation(
  data,
  censored_variable,
  censoring_indicator,
  covariates = NULL,
  id = NULL
)
}
\description{
mean residual life imputation, aka conditional imputation
}
