% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotECDFsTopDECluster.R
\name{plotECDFsTopDECluster}
\alias{plotECDFsTopDECluster}
\title{Plot ECDFs for top DE cluster}
\usage{
plotECDFsTopDECluster(d_ecdfs, path = ".",
  filename = "plot_ECDFs_top_DE_cluster.pdf")
}
\arguments{
\item{d_ecdfs}{ECDFs object from \code{\link{calcECDFs}}.}

\item{path}{Path to save plot. Default = current working directory.}

\item{filename}{Filename for plot.}
}
\description{
Basic diagnostic plot showing ECDFs (one per sample) for top differentially expressed 
(DE) cluster
}
\details{
Saves a plot showing ECDFs (one per sample) for the top differentially expressed (DE) 
cluster detected.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
