% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMST.R
\name{plotMST}
\alias{plotMST}
\title{Plot minimum spanning tree (MST)}
\usage{
plotMST(d_se, d_counts, res_DA = NULL, type = c("basic", "DA", "DE"),
  path = ".")
}
\arguments{
\item{d_se}{Data object from \code{\link{generateClusters}}, which contains the minimum
spanning tree (MST) information from the FlowSOM clustering algorithm.}

\item{d_counts}{Data object from \code{\link{calcCounts}}, which contains the number of
cells per cluster.}

\item{res_DA}{Results object from \code{\link{testDA}}, which contains the DA testing
results. Required if \code{type = DA}. Default = NULL.}

\item{type}{Type of MST plot to generate. Options are "basic" (no differential testing
results), "DA" (shading to represent DA testing results), "DE" (shading to represent
DE testing results). Default = "basic".}

\item{path}{Path to save plot. Default = current working directory.}
}
\description{
Plot minimum spanning tree (MST) from FlowSOM clustering, with optional color shading 
to represent differential testing results
}
\details{
Creates a minimum spanning tree (MST) plot, which shows the relationship between
clusters generated by FlowSOM.

The \code{type} argument specifies the type of MST plot to generate. In each case,
point sizes are also scaled to represent the total number of cells per cluster. Options
are:
\itemize{
\item \code{basic} Basic MST plot without any differential testing results.
\item \code{DA} MST plot with shading to represent results from tests for differential
abundance (DA) of clusters.
\item \code{DE} MST plot with shading to represent results from tests for differential
expression (DE) of functional markers within clusters.
}
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA, testDE_KS,
# testDE_LM

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
  \code{\link{testDE_KS}} \code{\link{testDE_LM}}
}
