% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopDACluster.R
\name{plotTopDACluster}
\alias{plotTopDACluster}
\title{Plot abundances for top DA cluster}
\usage{
plotTopDACluster(f1, filename = "./plot_top_DA_cluster.pdf")
}
\arguments{
\item{f1}{Output from \code{\link{testDA}}.}

\item{filename}{Filename and path to save plot. Default path is current working
directory.}
}
\description{
Basic diagnostic plot showing abundances for top differentially abundant (DA) cluster
}
\details{
Saves a plot showing abundances for the top differentially abundant (DA) cluster 
detected by \code{\link{testDA}}.
}
\examples{
# need to create a small example data set for examples
}
