% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopDAClusters.R
\name{plotTopDAClusters}
\alias{plotTopDAClusters}
\title{Plot abundances for top DA cluster}
\usage{
plotTopDAClusters(res_DA, path = ".", filename = "plot_top_DA_clusters.pdf")
}
\arguments{
\item{res_DA}{Fitted model objects from differential abundance tests with
\code{\link{testDA}}.}

\item{path}{Path to save plot. Default = current working directory.}

\item{filename}{Filename for plot.}
}
\description{
Basic diagnostic plot showing abundances for top differentially abundant (DA) cluster
}
\details{
Saves a plot showing abundances for the top differentially abundant (DA) cluster 
detected by \code{\link{testDA}}.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
