% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_mean_matching.R
\name{predictive_mean_matching}
\alias{predictive_mean_matching}
\title{Predictive mean matching analysis}
\usage{
predictive_mean_matching(
  data,
  censored_variable,
  censoring_indicator,
  variables_for_imputation = NULL,
  formula = NULL,
  regression_type = c("lm", "glm", "glmer"),
  repetitions = 10,
  weights = "weights",
  family = NULL
)
}
\arguments{
\item{data}{'data.frame'}

\item{censored_variable}{name of column containing censored data}

\item{censoring_indicator}{name of column containing indication if observed
(1) or censored (0) value in column 'censored_variable'}

\item{variables_for_imputation}{name of variable to use for imputation}

\item{formula}{the formula for fitting the regression model (e.g. formula(y~x))}

\item{regression_type}{function. The regression type to be used, lm for linear
regression, glm for general linear regression, glmer for generalized
linear mixed-effects models. Default: lm}

\item{repetitions}{number of repetitions for multiple imputation. Default: 10}

\item{weights}{name of column containing weights to be used in fitting the
regression model. Default = 'weights'. Ignored if no weights used.}

\item{family}{The family to be used in the regression model. Default = "binomial". Omited if linear model is used.}
}
\value{
list of elements 'data', 'betasMean' (mean regression coef of censored
 covariate), 'betasVar' (mean variance of regression coef of censored covariate),
 'fits' (regression fits)
}
\description{
Perform predictive mean matching by treating censored values as missing
}
\examples{
 lm_formula <- formula(Y ~ Surv(X,I) + Z)
 data <- simulate_data(100, lm_formula, type = "lm", n_levels_fixeff=2)
 pmm_out <- predictive_mean_matching(data,"X","I",c("Y","Z"),Y~X+Z)
}
