% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data}
\usage{
prepareData(d_input, sample_IDs, group_IDs, marker_cols, lineage_cols,
  functional_cols)
}
\arguments{
\item{d_input}{Input data. Must be a list or \code{\link[flowCore]{flowSet}}.}

\item{sample_IDs}{Vector of sample IDs.}

\item{group_IDs}{Vector of group IDs.}

\item{marker_cols}{Vector of indices of all markers.}

\item{lineage_cols}{Vector of indices of lineage markers.}

\item{functional_cols}{Vector of indices of functional markers.}
}
\value{
d_se Returns data as a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
  containing a single matrix of data in the \code{assays} slot, together with row 
  meta-data (sample and group IDs) and column meta-data (protein marker names, logical
  vectors for: markers, linage markers, functional markers).
}
\description{
Prepare data into format required by 'diffcyt' functions
}
\details{
Functions in the 'diffcyt' package assume that input data is provided as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}}, which contains a single
matrix of data, together with row and column meta-data.

This function accepts a list or \code{\link[flowCore]{flowSet}} as input, concatenates 
the data tables into a single matrix, and adds row and column meta-data.

Row meta-data contains sample labels and group membership labels. Column meta-data 
contains protein marker names, and logical entries indicating whether each column is 
(i) a marker, (ii) a lineage marker, and (iii) a functional marker.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
