% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetVals.R
\name{subsetVals}
\alias{subsetVals}
\title{Subset values for each cluster-sample combination}
\usage{
subsetVals(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster 
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = markers (non-clustering markers
  only). Each entry is a vector of expression values with length equal to the number of
  cells.
}
\description{
Subset values for each cluster-sample combination, required for differential expression
testing method 'diffcyt-KS'
}
\details{
Subsets the expression values for each cluster-sample combination, and returns as a new
\code{\link{SummarizedExperiment}} object. The values can then be provided to 
\code{\link{testDE_KS}} for differential testing with method 'diffcyt-KS' 
(Kolmogorov-Smirnov tests).

The expression values are returned in a new 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = clusters,
columns = samples, sheets ('assay' slots) = markers (non-clustering markers only). Note
that there is a separate table of values ('assay') for each marker, and each 'value' in
the tables consists of a vector of expression values.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA, testDE_KS,
# testDE_LM

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}} 
  \code{\link{testDE_KS}} \code{\link{testDE_LM}}
}
