% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA_GLMM.R
\name{testDA_GLMM}
\alias{testDA_GLMM}
\title{Test for differential abundance: method 'diffcyt-DA-GLMM'}
\usage{
testDA_GLMM(d_counts, group_IDs, contrast = NULL, batch_IDs = NULL,
  covariates = NULL, block_IDs = NULL, min_cells = 3,
  min_samples = NULL)
}
\arguments{
\item{d_counts}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster cell counts, from \code{\link{calcCounts}}.}

\item{group_IDs}{Vector or factor of group membership labels for each sample (e.g.
diseased vs. healthy, or treated vs. untreated). Vectors are converted to factors
internally. The user must ensure that this is identical to the \code{group_IDs}
provided previously to \code{\link{prepareData}}.}

\item{contrast}{Contrast specifying the differential comparison of interest for
testing. Each contrast should be in a separate row. If not provided, the default is
to compare the second vs. first level of \code{group_IDs}.}

\item{batch_IDs}{(Optional) Vector or factor of batch IDs. Batch effects are removed by
adding the batch IDs to the GLMM.}

\item{covariates}{(Optional) Numeric matrix of continuous covariates, to be added to
the GLMM in order to remove their effects.}

\item{block_IDs}{(Optional) Vector or factor of block IDs, for paired experimental
designs.}

\item{min_cells}{Filtering parameter. Default = 3. Clusters are kept for differential
testing if they have at least \code{min_cells} cells in at least \code{min_samples}
samples in at least one condition.}

\item{min_samples}{Filtering parameter. Default = \code{min(table(group_IDs)) - 1},
i.e. one less than the number of samples in the smallest group. Clusters are kept for
differential testing if they have at least \code{min_cells} cells in at least
\code{min_samples} samples in at least one condition.}
}
\value{
Returns a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} object,
  with differential test results stored in the \code{rowData} slot. Results include raw
  p-values and adjusted p-values, which can be used to rank clusters by evidence for
  differential abundance. The results can be accessed with the
  \code{\link[SummarizedExperiment]{rowData}} accessor function.
}
\description{
Calculate tests for differential abundance of clusters using method 'diffcyt-DA-GLMM'
}
\details{
Calculates tests for differential abundance of clusters, using generalized linear mixed
models (GLMMs) for each cluster. This methodology was originally developed and
described by Nowicka et al. (2017), \emph{F1000Research}.

For more details on the underlying statistical methodology, refer to the paper by
Nowicka et al. (2017), \emph{F1000Research}. The implementation here contains several
additional modifications. In particular, we use high-resolution clustering and report
results at the high-resolution cluster level, instead of relying on a manual
cluster-merging step. We also include a filtering step to remove high-resolution
clusters with very small numbers of cells, to improve power.

Filtering: Clusters are kept for differential testing if they have at least
\code{min_cells} cells in at least \code{min_samples} samples in at least one
condition. This removes clusters with very low cell counts across conditions, which
improves statistical power.
}
\examples{
# to do

}
\seealso{
to do
}
