% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA_limma.R
\name{testDA_limma}
\alias{testDA_limma}
\title{Test for differential abundance}
\usage{
testDA_limma(d_counts, group_IDs, contrast = NULL, batch_IDs = NULL,
  block_IDs = NULL, covariates = NULL, min_cells = 3,
  min_samples = NULL, path = ".")
}
\arguments{
\item{d_counts}{to do}

\item{group_IDs}{to do. User must ensure that this is identical to 'group_IDs' provided
in 'prepareData'.}

\item{contrast}{to do}

\item{batch_IDs}{to do}

\item{block_IDs}{to do}

\item{covariates}{to do}

\item{min_cells}{to do}

\item{min_samples}{to do}

\item{path}{to do}
}
\value{
returns
}
\description{
Calculate tests for differential abundance of clusters
}
\details{
method: limma-voom
}
\examples{
# to do

}
\seealso{
to do
}
