% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDE_FDA.R
\name{testDE_FDA}
\alias{testDE_FDA}
\title{Test for differential expression of functional markers (method: "diffcyt-FDA")}
\usage{
testDE_FDA(d_ecdfs, d_clus, group, use_weights = TRUE, n_perm = 5000,
  n_cores = NULL)
}
\arguments{
\item{d_ecdfs}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object 
containing empirical cumulative distribution functions (ECDFs), calculated with 
\code{\link{calcECDFs}}.}

\item{d_clus}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object 
containing cluster medians (median expression of functional markers) and cluster 
frequencies (number of cells), from \code{\link{calcMediansAndFreq}}. The cluster 
frequencies are used as weights for the permutation t-tests (the cluster medians are 
not used here).}

\item{group}{Factor containing group membership for each sample (for example, diseased 
vs. healthy), for differential comparisons and statistical tests.}

\item{use_weights}{Whether to include weights (per cluster-sample combination) for 
weighted permutation tests (see details below). Without weights, runtime is much 
faster, but result do not account for uncertainty due to different numbers of cells 
per sample (within each cluster). Default = TRUE.}

\item{n_perm}{Number of permutations to use for permutation testing. Default = 5000 
(i.e. minimum possible p-value = 0.0002).}

\item{n_cores}{Number of processor cores for parallelized evaluation. Default = all
available cores.}
}
\value{
Returns a data frame containing cluster labels, functional marker names, and 
  p-values summarizing the evidence for differential expression. Rows are sorted by 
  p-value (smallest first).
}
\description{
Calculate tests for differential expression of functional markers within clusters 
(method: "diffcyt-FDA")
}
\details{
The "diffcyt-FDA" methodology uses techniques from functional data analysis (FDA) to 
model the functional marker expression signals within each cluster. The differential 
expression tests compare the FDA-modeled expression profiles within clusters between 
samples in the two groups (e.g. diseased vs. healthy).

Permutation tests are used to calculate p-values.

By default, the tests are weighted by the number of cells per cluster-sample
combination, representing the relative uncertainty in calculating each curve (i.e. the
ECDF for each cluster-sample combination). Note that these are relative weights within
each cluster only. Weights can also be disabled for (much) faster runtime
(\code{weights = FALSE}), but then the results will not account for this source of
uncertainty.

We use the \code{\link[fda]{fda}} package (Ramsay et al. 2014) for the FDA modeling 
steps; with modifications to the permutation t-testing function to allow weights and 
improve runtime.

The \code{\link[BiocParallel]{BiocParallel}} package is used for parallelized 
evaluation on multi-core systems, to speed up runtime.

Alternative methodologies for testing for differential expression of functional markers
are available in the functions \code{testDE_med} and \code{testDE-KS}.
}
\examples{
library(flowCore)
library(limma)

# filenames
files <- list.files(system.file("extdata", package = "diffcyt"), 
                    pattern = "\\\\.fcs$", full.names = TRUE)
files_BCRXL <- files[grep("BCRXL", files)]
files_ref <- files[grep("ref", files)]

# load data
files_load <- c(files_BCRXL, files_ref)
d_input <- lapply(files_load, read.FCS, transformation = FALSE, truncate_max_range = FALSE)

# sample IDs and group IDs
sample_IDs <- gsub("\\\\.fcs$", "", basename(files_load))
sample_IDs
group_IDs <- gsub("^patient[0-9]_", "", sample_IDs)
group_IDs

# indices of all marker columns, lineage markers, and functional markers
# (see Table 1 in Bruggner et al. 2014)
marker_cols <- c(3:4, 7:9, 11:19, 21:22, 24:26, 28:31, 33)
lineage_cols <- c(3:4, 9, 11,12,14, 21, 29, 31, 33)
functional_cols <- setdiff(marker_cols, lineage_cols)

# prepare data
d_se <- prepareData(d_input, sample_IDs, group_IDs, marker_cols, lineage_cols, functional_cols)

# transform data
d_se <- transformData(d_se, cofactor = 5)

# generate clusters (note: using small 3x3 SOM grid for demonstration purposes)
d_se <- generateClusters(d_se, cols_to_use = lineage_cols, xdim = 3, ydim = 3, 
                         seed = 123, plot = FALSE)

# calculate cluster medians and frequencies
d_clus <- calcMediansAndFreq(d_se)

# calculate ECDFs
d_ecdfs <- calcECDFs(d_se)


#########################################################################################
# (3) Test for differential expression (DE) of functional markers: method "diffcyt-FDA" #
#########################################################################################

# re-level factor to use "ref" as base level
group <- factor(group_IDs, levels = c("ref", "BCRXL"))

# note: using no weights, small number of permutations, and single core for 
# demonstration purposes
res_DE_FDA <- testDE_FDA(d_ecdfs, d_clus, group, use_weights = FALSE, 
                         n_perm = 10, n_cores = 1)

# (note: this is a small example data set used for demonstration purposes only; results
# are not biologically meaningful)
head(res_DE_FDA)

}
