% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDE_med.R
\name{testDE_med}
\alias{testDE_med}
\title{Test for differential expression within clusters: method 'diffcyt-med'}
\usage{
testDE_med(d_se, d_counts, d_medians, design, contrast, block_IDs = NULL,
  min_cells = 3, min_samples = NULL, plot = TRUE, path = ".")
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing marker
information in column meta-data.}

\item{d_counts}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster cell counts, from \code{\link{calcCounts}}.}

\item{d_medians}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster medians (median expression of functional markers), from
\code{\link{calcMedians}}.}

\item{design}{Design matrix, created with \code{\link{createDesignMatrix}}. See
\code{\link{createDesignMatrix}} for details.}

\item{contrast}{Contrast matrix, created with \code{\link{createContrast}}. See
\code{\link{createContrast}} for details.}

\item{block_IDs}{(Optional) Vector or factor of block IDs (e.g. patient IDs) for paired
experimental designs, to be included as random effects. If provided, the block IDs
will be included as random effects using the \code{limma}
\code{\link[limma]{duplicateCorrelation}} methodology. Alternatively, block IDs can
be included as fixed effects in the design matrix (\code{\link{createDesignMatrix}}).
See details.}

\item{min_cells}{Filtering parameter. Default = 3. Clusters are kept for differential
testing if they have at least \code{min_cells} cells in at least \code{min_samples}
samples in at least one condition. In addition, clusters containing zero cells for
any sample are removed, due to the requirements of \code{voom}.}

\item{min_samples}{Filtering parameter. Default = \code{min(table(group_IDs)) - 1},
i.e. one less than the number of samples in the smallest group. Clusters are kept for
differential testing if they have at least \code{min_cells} cells in at least
\code{min_samples} samples in at least one condition. In addition, clusters
containing zero cells for any sample are removed, due to the requirements of
\code{voom}.}

\item{plot}{Whether to save diagnostic plots for the \code{limma}
\code{\link[limma]{voom}} transformations. Default = TRUE.}

\item{path}{Path for diagnostic plots. Default = current working directory.}
}
\value{
Returns a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} object,
  where rows = cluster-marker combinations, columns = samples. In the rows, clusters
  are repeated for each functional marker (i.e. the sheets or 'assays' from the
  previous \code{d_medians} object are stacked into a single matrix). Differential test
  results are stored in the 'rowData' slot. Results include raw p-values and adjusted
  p-values from the \code{\link[limma]{limma}} empirical Bayes moderated tests, which
  can be used to rank cluster-marker combinations by evidence for differential
  expression. The results can be accessed with the
  \code{\link[SummarizedExperiment]{rowData}} accessor function.
}
\description{
Calculate tests for differential expression of functional markers within clusters using
method 'diffcyt-med'.
}
\details{
This method uses median marker expression values to characterize the functional marker
expression profiles within each cluster (i.e. the signal of interest). In addition,
empirical Bayes methods are used to improve statistical power by sharing information on
variability (i.e. variance across samples for a single cluster) between clusters. The
\code{\link[limma]{limma}} package (Ritchie et al. 2015, \emph{Nucleic Acids Research})
is used to fit linear models and calculate empirical Bayes moderated tests. Since count
data are often heteroscedastic, we use the  \code{\link[limma]{voom}} method (Law et
al. 2014, \emph{Genome Biology}) to transform the raw cluster cell counts and estimate
observation-level weights to stabilize the mean-variance relationship. Diagnostic plots
are shown if \code{plot = TRUE}.

The experimental design must be specified using a design matrix, which can be created
with \code{\link{createDesignMatrix}}. Flexible experimental designs are possible,
including batch effects, continuous covariates, and paired designs. See
\code{\link{createDesignMatrix}} for more details.

For paired designs, either fixed effects or random effects can be used. Fixed effects
are simpler, but random effects may improve power for unbalanced designs. To use fixed
effects, provide the block IDs (e.g. patient IDs) to \code{\link{createDesignMatrix}}.
To use random effects, provide the \code{block_IDs} argument here instead. This will
make use of the \code{limma} \code{\link[limma]{duplicateCorrelation}} methodology;
note that block IDs should not be included in the design matrix in this case.

The contrast matrix specifying the contrast of interest can be created with
\code{\link{createContrast}}. See \code{\link{createContrast}} for more details.

Filtering: Clusters are kept for differential testing if they have at least
\code{min_cells} cells in at least \code{min_samples} samples in at least one
condition. This removes clusters with very low cell counts across conditions, which
improves power. In addition, clusters containing zero cells for any sample are removed,
due to the requirements of \code{voom}.
}
\examples{
# to do

}
