% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{transformData}
\alias{transformData}
\title{Transform data}
\usage{
transformData(d_input, cofactor = 5, marker_cols = NULL)
}
\arguments{
\item{d_input}{Input data. Assumed to be in the form of a 
\code{\link[flowCore]{flowSet}} object. The input data format can be checked using 
the function \code{\link{checkInputData}}.}

\item{cofactor}{Cofactor for 'arcsinh' transform. Default = 5, which is appropriate 
for mass cytometry (CyTOF) data. For fluorescence flow cytometry, we recommend 
cofactor = 150 instead.}

\item{marker_cols}{Indices of protein marker columns. The transform will be applied to
these columns only. Default = NULL, which transforms all columns.}
}
\value{
d_transf Data with transform applied to protein marker columns.
}
\description{
Transform data prior to clustering
}
\details{
Flow and mass cytometry data should be transformed prior to clustering. This function 
implements an 'arcsinh' transform with adjustable 'cofactor' parameter. Recommended
values for the cofactor are 5 (mass cytometry, CyTOF) or 150 (fluorescence flow
cytometry); see Bendall et al. (2011), \emph{Science}, Supplementary Figure S2.
}
\examples{
# need to create a small example data set for examples
}
