% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCounts.R
\name{calcCounts}
\alias{calcCounts}
\title{Calculate cluster counts (frequencies)}
\usage{
calcCounts(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, assay = values (counts).
}
\description{
Calculate number of cells per cluster per sample (i.e. cluster counts / frequencies / 
abundances)
}
\details{
Calculate number of cells per cluster per sample (referred to as cluster 'counts', 
'frequencies', or 'abundances').

The cluster counts are used as weights in the subsequent statistical tests and 
calculation of false discovery rates (FDRs).

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} 
object, where rows = clusters, columns = samples, assay = values (counts).
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
