% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCounts.R
\name{calcCounts}
\alias{calcCounts}
\title{Calculate cluster cell counts}
\usage{
calcCounts(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, assay = values (counts).
}
\description{
Calculate number of cells per cluster-sample combination
}
\details{
Calculate number of cells per cluster-sample combination (referred to as cluster cell
'counts', 'abundances', or 'frequencies').

The cluster cell counts are required for the differential abundance tests, and are also
used as weights to improve statistical power for the differential functional state
tests.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = samples, assay = values (counts). (Note that
this structure differs from the input data object, where rows = cells, and columns =
markers.)
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
