% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcECDFs.R
\name{calcECDFs}
\alias{calcECDFs}
\title{Calculate empirical cumulative distribution functions (ECDFs)}
\usage{
calcECDFs(d_se, resolution = 30)
}
\arguments{
\item{d_se}{Data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster 
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}

\item{resolution}{Resolution for evaluating ECDFs. The value of each ECDF is calculated
at this number of equally spaced points between the maximum and minimum observed 
marker expression values for a given cluster-sample combination. Default = 30.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = functional markers. Each entry 
  is a vector of values (i.e. the evaluated ECDF values) with length \code{resolution}.
}
\description{
Calculate empirical cumulative distribution functions (ECDFs) required for
methods "diffcyt-FDA", "diffcyt-KS", and "diffcyt-LM"
}
\details{
Methods "diffcyt-FDA", "diffcyt-KS", and "diffcyt-LM" use the empirical cumulative
distribution functions (ECDFs) of functional marker expression profiles (for each
cluster-sample combination) to test for differential expression of functional markers
between sample groups.

This function calculates the ECDFs for each cluster-sample combination for each 
functional marker, and evaluates each of them at a set of equally-spaced points. The 
ECDF values can then be used in the tests for differential expression.

The ECDF values are returned in a new
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = clusters,
columns = samples, sheets ('assay' slots) = functional markers. Note that each "value"
is a vector of evaluated ECDF values.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
