% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcECDFs.R
\name{calcECDFs}
\alias{calcECDFs}
\title{Calculate empirical cumulative distribution functions (ECDFs)}
\usage{
calcECDFs(d_se, resolution = 30)
}
\arguments{
\item{d_se}{Transformed data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, with cluster labels 
added in row meta-data using \code{\link{generateClusters}}.}

\item{resolution}{Resolution for evaluating ECDFs. The value of each ECDF is calculated
at this number of equally spaced points between the maximum and minimum observed
marker expression values for a given cluster-sample combination. Default = 30.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = functional markers. Each entry
  is a list of values (i.e. the evaluated ECDF values) with length \code{resolution}.
}
\description{
Calculate empirical cumulative distribution functions (ECDFs) for FDA-based methods
}
\details{
The functional data analysis (FDA) based methodology in "diffcyt-FDA" uses empirical
cumulative distribution functions (ECDFs) of the functional marker expression profiles 
by cluster and sample, to test for differential expression between sample groups.

This function calculates the ECDFs for each cluster-sample combination for each 
functional marker, and evaluates each of them at a set of equally-spaced points. These 
ECDF values can then be used for differential testing with \code{testDE_FDA}.

The ECDF values are returned in a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} with the same shape as the
output from \code{\link{calcMediansAndFreq}}: rows = clusters, columns = samples,
sheets ('assay' slots) = functional markers.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
