% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcECDFs.R
\name{calcECDFs}
\alias{calcECDFs}
\title{Calculate empirical cumulative distribution functions (ECDFs)}
\usage{
calcECDFs(d_se, resolution = 30)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}). Column
meta-data is assumed to contain vectors \code{is_marker_col}, \code{is_identity_col},
and \code{is_func_col}.}

\item{resolution}{Resolution for evaluating ECDFs. The value of each ECDF curve is
calculated at this number of equally spaced points between the minimum and maximum
observed (transformed) marker expression values for each cluster-sample combination.
Default = 30.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, sheets ('assay' slots) = markers. Each entry is a vector
  of values (i.e. the evaluated ECDF values) with length \code{resolution}. The
  \code{metadata} slot contains variables \code{id_identity_markers} and
  \code{id_func_markers}, which can be accessed with
  \code{metadata(d_medians)$id_identity_markers} and
  \code{metadata(d_medians)$id_func_markers}.
}
\description{
Calculate empirical cumulative distribution functions (ECDFs) required for differential
functional state testing
}
\details{
Empirical cumulative distribution functions (ECDFs) of marker expression profiles are
used to represent the signal for each cluster-sample combination. The ECDFs for the
functional markers can then be used to test for differential functional states of
clusters between sample groups.

By using the ECDFs, these methods take the full distributions of the functional marker
expression profiles into account. By contrast, most existing approaches rely on only a
single feature value or summary statistic for each marker expression profile (such as
the median), which discards a significant amount of extra information contained in the
rest of the distributions.

This function calculates the ECDFs for each cluster-sample combination for each marker,
and evaluates them at a set of equally-spaced points. The evaluated ECDF values for the
functional markers can then be used in the differential functional state tests.

The input data object is assumed to contain vectors \code{is_marker_col},
\code{is_identity_col}, and \code{is_func_col} in the column meta-data (see
\code{\link{prepareData}}). These indicate the sets of all marker columns, identity
marker columns, and functional marker columns.

The evaluated ECDF values are returned in a new
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = clusters,
columns = samples, sheets ('assay' slots) = markers. Note that there is a separate
table of values ('assay') for each marker, and each 'value' in the tables consists of a
vector of evaluated ECDF values. The \code{metadata} slot also contains variables
\code{id_identity_markers} and \code{id_func_markers}, which can be used to identify
the sets of identity and functional markers.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
