% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMedians.R
\name{calcMedians}
\alias{calcMedians}
\title{Calculate cluster medians}
\usage{
calcMedians(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, sheets ('assay' slots) = markers.
}
\description{
Calculate cluster medians (median expression of each marker for each cluster-sample
combination)
}
\details{
Calculate median expression of each marker for each cluster-sample combination.

The data object is assumed to contain a vector \code{is_marker_col} in the column
meta-data (see \code{\link{prepareData}}), which indicates whether each column
represents a marker. Cluster medians are calculated for these columns.

The cluster medians (by sample) are required for differential expression tests.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = samples, sheets ('assay' slots) = markers.
Note that there is a separate table of values ('assay') for each marker. The set of
functional markers can be identified using the variable 'id_func_markers' saved in the
'metadata' slot.
}
\examples{
# See full examples in testing functions.

}
\seealso{
to do
}
