% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMedians.R
\name{calcMedians}
\alias{calcMedians}
\title{Calculate cluster medians}
\usage{
calcMedians(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster 
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = markers (non-clustering markers
  only).
}
\description{
Calculate cluster medians (median expression of each non-clustering marker for each 
cluster-sample combination)
}
\details{
Calculate median expression of each non-clustering marker for each cluster-sample
combination.

The data object is assumed to contain a vector \code{is_DE_col} in the column meta-data
(see \code{\link{prepareData}}), which indicates whether each column is a 
'non-clustering' marker to be used for differential expression analysis. Cluster 
medians are calculated for these markers only.

The cluster medians are used for the differential expression tests, and for
visualization of results.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} 
object, where rows = clusters, columns = samples, sheets ('assay' slots) = markers 
(non-clustering markers only). Note that there is a separate table of values ('assay') 
for each marker.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA, testDE_KS,
# testDE_LM

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
  \code{\link{testDE_KS}} \code{\link{testDE_LM}}
}
