% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMedians.R
\name{calcMedians}
\alias{calcMedians}
\title{Calculate cluster medians}
\usage{
calcMedians(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster 
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = 
  clusters, columns = samples, sheets ('assay' slots) = functional markers.
}
\description{
Calculate cluster medians (median functional marker expression per cluster per sample)
}
\details{
Calculate median expression of functional markers (cluster medians) per cluster per
sample; i.e. one table of values per functional marker; each containing one value for 
each cluster in each sample.

The cluster medians are used in the subsequent statistical tests for method 
"diffcyt-med", as well as for visualization of results from all methods.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} 
object, where rows = clusters, columns = samples, sheets ('assay' slots) = functional 
markers.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
