% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMedians.R
\name{calcMedians}
\alias{calcMedians}
\title{Calculate cluster medians (by sample)}
\usage{
calcMedians(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}). Column
meta-data is assumed to contain vectors \code{is_marker}, \code{is_type_marker}, and
\code{is_state_marker}.}
}
\value{
\code{d_medians}: \code{\link[SummarizedExperiment]{SummarizedExperiment}}
  object, where rows = clusters, columns = samples, sheets ('assay' slots) = markers.
  The \code{metadata} slot contains variables \code{id_type_markers} and
  \code{id_state_markers}, which can be accessed with
  \code{metadata(d_medians)$id_type_markers} and
  \code{metadata(d_medians)$id_state_markers}.
}
\description{
Calculate cluster medians by sample (median expression of each marker for each
cluster-sample combination)
}
\details{
Calculate median expression of each marker by sample, for each cluster-sample
combination.

The data object is assumed to contain vectors \code{is_marker}, \code{is_type_marker},
and \code{is_state_marker} in the column meta-data (see \code{\link{prepareData}}).
These indicate the sets of all marker columns, cell type marker columns, and cell state
marker columns. Cluster medians are calculated for all markers.

The cluster medians (by sample) are required for testing for differential states within
cell populations, and for plotting.

Variables \code{id_type_markers} and \code{id_state_markers} are saved in the
\code{metadata} slot of the output object. These can be used to identify the cell type
and cell state markers in later steps of the 'diffcyt' pipeline.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = samples, sheets ('assay' slots) = markers.
Note that there is a separate table of values ('assay') for each marker. The
\code{metadata} slot also contains variables \code{id_type_markers} and
\code{id_state_markers}, which can be used to identify the sets of cell type and cell
state markers.
}
\examples{
# For a full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline, see the package vignette.

# Create some random data (without differential signal)
cofactor <- 5
set.seed(123)
d_input <- list(
  sample1 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample2 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample3 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample4 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor
)

sample_info <- data.frame(
  sample = factor(paste0("sample", 1:4)), 
  group = factor(c("group1", "group1", "group2", "group2")), 
  stringsAsFactors = FALSE
)

marker_info <- data.frame(
  marker_name = paste0("marker", 1:20), 
  is_marker = rep(TRUE, 20), 
  is_type_marker = c(rep(TRUE, 10), rep(FALSE, 10)), 
  is_state_marker = c(rep(FALSE, 10), rep(TRUE, 10)), 
  stringsAsFactors = FALSE
)

# Prepare data
d_se <- prepareData(d_input, sample_info, marker_info)

# Transform data
d_se <- transformData(d_se)

# Generate clusters
d_se <- generateClusters(d_se)

# Calculate medians (by sample)
d_medians <- calcMedians(d_se)

}
