% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMediansAll.R
\name{calcMediansAll}
\alias{calcMediansAll}
\title{Calculate cluster medians across all samples (for plotting)}
\usage{
calcMediansAll(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}). Column
meta-data is assumed to contain vectors \code{is_marker}, \code{is_celltype_marker},
and \code{is_state_marker}.}
}
\value{
\code{d_medians_all}: \code{\link[SummarizedExperiment]{SummarizedExperiment}}
  object, where rows = clusters, columns = markers, assay = values (marker expression
  values). The \code{metadata} slot contains variables \code{id_celltype_markers} and
  \code{id_state_markers}, which can be accessed with
  \code{metadata(d_medians)$id_celltype_markers} and
  \code{metadata(d_medians)$id_state_markers}.
}
\description{
Calculate cluster medians (median expression of each marker) for each cluster
}
\details{
Calculate median expression of each marker for each cluster across all samples.

The data object is assumed to contain vectors \code{is_marker},
\code{is_celltype_marker}, and \code{is_state_marker} in the column meta-data (see
\code{\link{prepareData}}). These indicate the sets of all marker columns, cell type
marker columns, and functional state marker columns. Cluster medians are calculated for
all markers.

The cluster medians (across all samples) are required for plotting functions.

Variables \code{id_celltype_markers} and \code{id_state_markers} are saved in the
\code{metadata} slot of the output object. These can be used to identify the cell type
and functional state markers in later steps of the 'diffcyt' pipeline.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = markers, assay = values (marker expression
values). The \code{metadata} slot also contains variables \code{id_celltype_markers}
and \code{id_state_markers}, which can be used to identify the sets of cell type and
functional state markers.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
