% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMediansAll.R
\name{calcMediansAll}
\alias{calcMediansAll}
\title{Calculate cluster medians across all samples (for plotting)}
\usage{
calcMediansAll(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}). Column
meta-data is assumed to contain vectors \code{is_marker_col}, \code{is_identity_col},
and \code{is_func_col}.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = markers, assay = values (marker expression values). The
  \code{metadata} slot contains variables \code{id_identity_markers} and
  \code{id_func_markers}, which can be accessed with
  \code{metadata(d_medians)$id_identity_markers} and
  \code{metadata(d_medians)$id_func_markers}.
}
\description{
Calculate cluster medians (median expression of each marker) for each cluster
}
\details{
Calculate median expression of each marker for each cluster across all samples.

The data object is assumed to contain vectors \code{is_marker_col},
\code{is_identity_col}, and \code{is_func_col} in the column meta-data (see
\code{\link{prepareData}}). These indicate the sets of all marker columns, identity
marker columns, and functional marker columns. Cluster medians are calculated for all
markers.

The cluster medians (across all samples) are required for plotting functions.

Variables \code{id_identity_markers} and \code{id_func_markers} are saved in the
\code{metadata} slot of the output object. These can be used to identify the identity
and functional markers in later steps of the 'diffcyt' pipeline.

Results are returned as a new \code{\link[SummarizedExperiment]{SummarizedExperiment}}
object, where rows = clusters, columns = markers, assay = values (marker expression
values). The \code{metadata} slot also contains variables \code{id_identity_markers}
and \code{id_func_markers}, which can be used to identify the sets of identity and
functional markers.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
