% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSubsetVals.R
\name{calcSubsetVals}
\alias{calcSubsetVals}
\title{Subset values for each cluster-sample combination}
\usage{
calcSubsetVals(d_se)
}
\arguments{
\item{d_se}{Data object from previous steps, in
\code{\link[SummarizedExperiment]{SummarizedExperiment}} format, containing cluster
labels as a column in the row meta-data (from \code{\link{generateClusters}}).}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows =
  clusters, columns = samples, sheets ('assay' slots) = markers. Each entry is a vector
  of expression values with length equal to the number of cells.
}
\description{
Subset values for each cluster-sample combination for each marker, required for
differential expression testing
}
\details{
Subsets the expression values for each cluster-sample combination for each marker, and
returns as a new \code{\link{SummarizedExperiment}} object. The values can then be
provided to the differential expression testing functions.

The expression values are returned in a new
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, where rows = clusters,
columns = samples, sheets ('assay' slots) = markers. Note that there is a separate
table of values ('assay') for each marker, and each 'value' in the tables consists of a
vector of expression values.
}
\examples{
# See full examples in testing functions.

}
\seealso{
to do
}
