% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrast.R
\name{createContrast}
\alias{createContrast}
\title{Create contrast matrix}
\usage{
createContrast(group_IDs, contrast = NULL)
}
\arguments{
\item{group_IDs}{Vector or factor of group membership labels for each sample (e.g.
diseased vs. healthy, or treated vs. untreated). Vectors are converted to factors
internally. The first level of the factor (e.g. healthy) or first entry of the vector
will be used as the reference level for differential testing. To re-order factor
levels, use \code{\link[stats]{relevel}} or \code{\link[base]{factor}}.}

\item{contrast}{Vector defining the contrast of interest. For example, \code{c(0, 1)}
to compare the second level against the first level for a \code{group_IDs} vector of
length 2. See details for additional explanation. Default is to compare the second
level against the first level of \code{group_IDs}.}
}
\value{
Returns a contrast matrix containing the contrast of interest, formatted as a
  matrix with a single column.
}
\description{
Create contrast matrix for differential testing
}
\details{
Creates a contrast matrix in the correct format for differential testing. This can then
be provided to the differential testing functions, together with either a design matrix
or model formula, and the data object.

The argument \code{contrast} defines the contrast of interest for the levels of
\code{group_IDs}. Note that the model formula (\code{\link{createFormula}}) or design
matrix (\code{\link{createDesignMatrix}}) assumes that the first level of
\code{group_IDs} is the reference level, and the model formula or design matrix are
specified so that the first column of the design matrix is an intercept. In this setup,
the subsequent terms in the design matrix represent the difference compared to the
reference.

For example, to compare the second level against the first level for a \code{group_IDs}
vector of length 2, the \code{contrast} argument would be \code{c(0, 1)}. To compare
the second level against the first level for a \code{group_IDs} vector of length 3, it
would be \code{c(0, 1, 0)}.

The contrast matrix is formatted as a matrix with a single column containing the
contrast of interest. To perform tests for multiple contrasts, run this function and
the corresponding differential testing function multiple times.
}
\examples{
# to do

}
\seealso{
to do
}
