% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{createFormula}
\alias{createFormula}
\title{Create model formula and data frame of variables}
\usage{
createFormula(group_IDs, batch_IDs = NULL, covariates = NULL,
  block_IDs = NULL, block_IDs_type = c("fixed", "random"),
  sample_IDs = NULL)
}
\arguments{
\item{group_IDs}{Vector or factor of group membership labels for each sample (e.g.
diseased vs. healthy, or treated vs. untreated). These are the parameters of interest
in the model. Vectors are converted to factors internally. The first level of the
factor (e.g. healthy) or first entry of the vector will be used as the reference
level for differential testing. To re-order factor levels, use
\code{\link[stats]{relevel}} or \code{\link[base]{factor}}.}

\item{batch_IDs}{(Optional) Vector or factor of batch IDs. Batch IDs are included as
fixed effects in the model formula. This allows batch effects to be estimated during
model fitting and taken into account during inference on the parameters of interest.}

\item{covariates}{(Optional) Numeric matrix of continuous covariates. Covariates are
included as fixed effects in the model formula. This allows their effects to be
estimated during model fitting and taken into account during inference on the
parameters of interest.}

\item{block_IDs}{(Optional) Vector or factor of block IDs, e.g. for paired designs
(e.g. one diseased and one healthy sample per patient). Block IDs can be included as
either fixed effects or random intercept terms; this choice is specified using the
argument \code{block_IDs_type}. Note that some testing methods (e.g.
'diffcyt-DA-limma') require random intercept block IDs to be provided directly
instead.}

\item{block_IDs_type}{(Optional) Whether block IDs should be included in the model
formula as fixed effects or random effects. Options are 'fixed' and 'random'. Default
= 'fixed'.}

\item{sample_IDs}{(Optional) Vector or factor of sample IDs. Sample IDs are included as
random intercepts in the model formula, to account for overdispersion typically seen
in high-dimensional cytometry data. This is also known as an 'observation-level
random effect'; for more details see Nowicka et al. (2017), \emph{F1000Research}.}
}
\value{
Returns a list with two elements:
\itemize{
\item \code{formula}: model formula
\item \code{data}: data frame of variables corresponding to the model formula
}
}
\description{
Create model formula and corresponding data frame of variables for model fitting
}
\details{
Creates a model formula and corresponding data frame of variables specifying the models
to be fitted. Similar to \code{\link{createDesignMatrix}}, but returns a model formula
instead of a design matrix.

The output is a list containing the model formula and corresponding data frame (one
column per formula term). These can then be provided to differential testing functions
that require a model formula, together with the main data object and contrast matrix.
}
\examples{
# to do

}
\seealso{
to do
}
