% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_generateClusters.R
\name{generateClusters}
\alias{generateClusters}
\title{Generate clusters}
\usage{
generateClusters(d_se, cols_to_use = NULL, xdim = 30, ydim = 30,
  meta_clustering = FALSE, meta_k = 40, seed = NULL, ...)
}
\arguments{
\item{d_se}{Transformed input data, from \code{\link{prepareData}} and
\code{\link{transformData}}.}

\item{cols_to_use}{Columns to use for clustering. Default = \code{NULL}, in which case
the markers specified by \code{is_clustering_col} in the column meta-data of
\code{d_se} will be used.}

\item{xdim}{Width of grid for self-organizing map for FlowSOM clustering (number of
clusters = \code{xdim} * \code{ydim}). Default = 30 (i.e. 900 clusters).}

\item{ydim}{Height of grid for self-organizing map for FlowSOM clustering (number of
clusters = \code{xdim} * \code{ydim}). Default = 30 (i.e. 900 clusters).}

\item{meta_clustering}{Whether to include FlowSOM 'meta-clustering' step. Default =
\code{FALSE}.}

\item{meta_k}{Number of meta-clusters for FlowSOM, if \code{meta-clustering = TRUE}.
Default = 40.}

\item{seed}{Random seed. Set to an integer value to generate reproducible results.
Default = \code{NULL}.}

\item{...}{Other parameters to pass to the FlowSOM clustering algorithm (through the
function \code{\link[FlowSOM]{BuildSOM}}).}
}
\value{
Returns the \code{\link[SummarizedExperiment]{SummarizedExperiment}} input
  object, with cluster labels for each cell stored in an additional column of row
  meta-data. Row meta-data can be accessed with
  \code{\link[SummarizedExperiment]{rowData}}. The minimum spanning tree (MST) object
  is stored in the experiment \code{metadata} slot, and can be accessed with
  \code{metadata(d_se)$MST}.
}
\description{
Generate clusters for \code{diffcyt} analysis
}
\details{
Performs clustering to divide cells into clusters representing populations or subsets,
which can then be further analyzed for differential abundance, or differential
expression within clusters. By default, a large number of small clusters is generated
(i.e. over-clustering), to ensure that rare populations are adequately separated from
larger ones.

Data is assumed to be in the form of a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object generated with
\code{\link{prepareData}} and transformed with \code{\link{transformData}}.

The input data object \code{d_se} is assumed to contain a vector
\code{is_clustering_col} in the column meta-data. This vector specifies the columns
(protein markers) to be used for clustering. For example, in immunological data, this
may be the lineage markers. The choice of clustering markers is an important design
choice for the user, and will depend on the underlying experimental design and research
questions. It may be made based on prior biological knowledge or using data-driven
methods. For an example of a data-driven method of marker ranking and selection, see
Nowicka et al. (2017), \emph{F1000Research}.

We use the \code{\link[FlowSOM]{FlowSOM}} clustering algorithm (Van Gassen et al. 2015,
\emph{Cytometry Part A}, available from Bioconductor) to generate the clusters. We
previously showed that \code{FlowSOM} gives very good clustering performance for
high-dimensional cytometry data, for both major and rare cell populations, and is
extremely fast (Weber and Robinson, 2016, \emph{Cytometry Part A}).

By default, the clustering is run at high resolution to give a large number of small
clusters (over-clustering). This is done by running only the initial 'self-organizing
map' clustering step in the \code{FlowSOM} algorithm, i.e. without the final
'meta-clustering' step. This ensures that small or rare populations are adequately
separated from larger populations, which is crucial for detecting differential
abundance or differential expression signals from rare populations.

The minimum spanning tree (MST) object from \code{\link[FlowSOM]{BuildMST}} (required
for MST plots) is stored in the experiment \code{metadata} slot in the
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object \code{d_se}, and can be
accessed with \code{metadata(d_se)$MST}.
}
\examples{
# See full examples in testing functions.

}
\seealso{
to do
}
