% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_processing.R
\name{is_valid_censored_formula}
\alias{is_valid_censored_formula}
\title{check if 'Surv()' formula is valid}
\usage{
is_valid_censored_formula(formula, throw_error = TRUE)
}
\arguments{
\item{formula}{formula containing indication 'Surv()'.
See \code{\link{testDA_censoredGLMM}}}

\item{throw_error}{should an error be thrown if 'formula' is incorrect.
default = 'TRUE'.}
}
\value{
logical
}
\description{
check if 'Surv()' formula is valid
}
