% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMST.R
\name{plotMST}
\alias{plotMST}
\title{Plot minimum spanning tree (MST)}
\usage{
plotMST(d_se, path = ".", filename = "plot_MST.pdf", ...)
}
\arguments{
\item{d_se}{Data object in \code{\link[SummarizedExperiment]{SummarizedExperiment}} 
format. Must contain a column of cluster labels in the row meta-data, and a minimum 
spanning tree (MST) object in the experiment meta-data, as generated by 
\code{\link{generateClusters}}.}

\item{path}{Path to save plot. Default = "." (current working directory).}

\item{filename}{Filename to save plot.}

\item{...}{Additional arguments to pass to \code{plot}.}
}
\description{
Plot minimum spanning tree (MST) from FlowSOM clustering
}
\details{
Saves a minimum spanning tree (MST) plot, which shows the hierarchical relationship 
between clusters detected by FlowSOM. Point sizes are scaled to represent cluster sizes
(number of cells per cluster).
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
}
