% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMST.R
\name{plotMST}
\alias{plotMST}
\title{Plot minimum spanning trees (MST)}
\usage{
plotMST(d_se, d_counts, res_DA = NULL, res_DE = NULL, type = c("basic",
  "DA", "DE"), path = ".")
}
\arguments{
\item{d_se}{Data object from \code{\link{generateClusters}}, which contains the minimum
spanning tree (MST) information from the FlowSOM clustering algorithm.}

\item{d_counts}{Data object from \code{\link{calcCounts}}, which contains the number of
cells per cluster.}

\item{res_DA}{Results object from the differential abundance (DA) testing function
(\code{\link{testDA}}). Required if \code{type = DA}. Default = NULL.}

\item{res_DE}{Results object from one of the differential expression (DE) testing
functions (\code{\link{testDE_med}}, \code{\link{testDE_FDA}},
\code{\link{testDE_KS}}, or \code{\link{testDE_LM}}). Required if \code{type = DE}.
Default = NULL.}

\item{type}{Type of MST plot to generate. Options are "basic" (no differential test 
results), "DA" (shading to represent DA test results), "DE" (shading to represent DE
test results). Default = "basic".}

\item{path}{Path to save plot. Default = current working directory.}
}
\description{
Plot minimum spanning trees (MST) from FlowSOM clustering, with optional color shading 
to represent differential test results
}
\details{
Creates one or more minimum spanning tree (MST) plots, which show the relationship
between clusters generated by FlowSOM. Optionally, points can be shaded to represent 
differential test results.

The \code{type} argument specifies the type of MST plots to generate. Options are:
\itemize{
\item \code{basic} Basic MST plot without any differential test results.
\item \code{DA} MST plot with shading to represent results from tests for differential 
abundance (DA) of clusters.
\item \code{DE} MST plots with shading to represent results from tests for differential
expression (DE) of functional markers within clusters. One MST plot is generated for 
each marker, displayed on multiple panels.
}

In each case, point sizes are also scaled to represent the total number of cells per 
cluster (across all samples).
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA, testDE_KS,
# testDE_LM

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
  \code{\link{testDE_KS}} \code{\link{testDE_LM}}
}
