% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMST.R
\name{plotMST}
\alias{plotMST}
\title{Plot minimum spanning tree (MST)}
\usage{
plotMST(d_se, d_counts, path = ".")
}
\arguments{
\item{d_se}{Data object from \code{\link{generateClusters}}, which contains the minimum
spanning tree (MST) information from the FlowSOM clustering algorithm.}

\item{d_counts}{Data object from \code{\link{calcCounts}}, which contains the number of
cells per cluster.}

\item{path}{Path to save plot. Default = current working directory.}
}
\description{
Plot minimum spanning tree (MST) from FlowSOM clustering
}
\details{
Creates a minimum spanning tree (MST) plot showing the relationship between clusters 
generated by FlowSOM.

This function generates a basic MST plot. Cluster sizes are scaled by the total number 
of cells per cluster.

Additional functions are available for MST plots that also display results from the 
differential expression tests.
}
\examples{
# See full examples in testing functions: testDA, testDE_med, testDE_FDA, testDE_KS,
# testDE_LM

}
\seealso{
\code{\link{testDA}} \code{\link{testDE_med}} \code{\link{testDE_FDA}}
  \code{\link{testDE_KS}} \code{\link{testDE_LM}}
}
