% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data}
\usage{
prepareData(d_input, sample_IDs, group_IDs, cols_markers = NULL,
  cols_type = NULL, cols_state = NULL, col_names = NULL,
  subsampling = FALSE, n_sub = NULL)
}
\arguments{
\item{d_input}{Input data. Must be a list or \code{\link[flowCore]{flowSet}} (one list
item or \code{\link[flowCore]{flowFrame}} per sample).}

\item{sample_IDs}{Vector of sample IDs.}

\item{group_IDs}{Vector of group IDs.}

\item{cols_markers}{Column indices indicating all protein markers.}

\item{cols_type}{Column indices indicating cell type markers, to be used for
clustering and testing for differential abundance.}

\item{cols_state}{Column indices indicating functional state markers, to be used for
testing for differential functional states.}

\item{col_names}{Optional vector of column names; for example, this is useful if the
column names in the input data contain channel names instead of marker names. Default
= column names of input data.}

\item{subsampling}{Whether to use random subsampling to select an equal number of cells
from each sample. Default = FALSE.}

\item{n_sub}{Number of cells to select from each sample by random subsampling, if
\code{subsampling = TRUE}. Default = number of cells in smallest sample.}
}
\value{
d_se Returns data as a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
  containing a single matrix of data (expression values) in the \code{assays} slot,
  together with row meta-data (sample IDs, group IDs) and column meta-data (protein
  marker names, logical vectors for: all markers, cell type markers, and functional
  state markers). The \code{group_IDs} vector is also stored in the \code{metadata}
  slot, and can be accessed with \code{metadata(d_se)$group_IDs}.
}
\description{
Prepare data into format required for \code{diffcyt} pipeline
}
\details{
Functions in the \code{diffcyt} analysis pipeline assume that input data is provided as
a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object, which contains a
single matrix of expression values, together with row and column meta-data.

This function accepts a list or \code{\link[flowCore]{flowSet}} as input (containing
one list item or \code{\link[flowCore]{flowFrame}} per sample), concatenates the data
tables into a single matrix, and adds row and column meta-data.

Row meta-data contains sample labels and group membership labels. Column meta-data
contains protein marker names, and logical entries indicating whether each column is
(i) a marker, (ii) a cell type marker (for clustering and testing for differential
abundance), and (iii) a functional state marker (for testing for differential
functional states).

Optionally, random subsampling can be used to select an equal number of cells from each
sample (\code{subsampling = TRUE}). This can be useful when there are large differences
in total numbers of cells per sample, since it ensures that samples with relatively
large numbers of cells do not dominate the clustering. However, subsampling should
generally not be used when rare cell populations are of interest, due to the
significant loss of information if cells from the rare population are discarded.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
