% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA.R
\name{testDA}
\alias{testDA}
\title{Test clusters for differential abundance}
\usage{
testDA(d_transf, clus, group, tbl_prop)
}
\arguments{
\item{d_transf}{Transformed input data from the previous steps. This should be in the
form of a \code{\link[flowCore]{flowSet}} object from the
\code{\link[flowCore]{flowCore}} package.}

\item{clus}{Cluster labels for individual cells, from \code{\link{generateClusters}}.}

\item{group}{Factor containing group membership for each sample (for example, diseased
vs. healthy). This is required to fit linear models and calculate statistical tests 
for each cluster. [to do: update this to extract it automatically from the flowSet 
object]}

\item{tbl_prop}{Table of cluster proportions (rows = clusters, columns = samples).
Calculated by function \code{\link{calculateFreq}}. [to do: update to save tbl_freq
and tbl_prop directly in an object, possibly as an additional sheet in a Summarized
Experiment]}
}
\value{
Returns \code{\link[limma]{limma}} fitted model objects for each cluster,
  including p-values from empirical Bayes moderated tests. The p-values can be
  accessed with \code{\link[limma]{topTable}} from the \code{limma} package.
}
\description{
Calculate empirical Bayes moderated tests for differential abundance of clusters
}
\details{
Calculates tests for differential abundance (differential frequencies) of clusters, 
using empirical Bayes moderation of cluster-wise variances to improve power.

We use the \code{\link[limma]{limma}} package (Ritchie et al. 2015, \emph{Nucleic
Acids Research}) to calculate the empirical Bayes moderated tests. Empirical Bayes
methods improve statistical power by sharing information on variability (i.e. variance
across samples for a single cluster) across clusters.

This function should be run after generating clusters with
\code{\link{generateClusters}}.
}
\examples{
# need to create a small example data set for examples
}
