% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDA_GLMM.R
\name{testDA_GLMM}
\alias{testDA_GLMM}
\title{Test for differential abundance: method 'diffcyt-DA-GLMM'}
\usage{
testDA_GLMM(d_counts, formula, contrast, min_cells = 3, min_samples = NULL)
}
\arguments{
\item{d_counts}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster cell counts, from \code{\link{calcCounts}}.}

\item{formula}{Model formula object, created with \code{\link{createFormula}}. This
should be a list containing two elements: \code{formula} and \code{data}; the model
formula and data frame of corresponding variables. See \code{\link{createFormula}}
for details.}

\item{contrast}{Contrast matrix, created with \code{\link{createContrast}}. See
\code{\link{createContrast}} for details.}

\item{min_cells}{Filtering parameter. Default = 3. Clusters are kept for differential
testing if they have at least \code{min_cells} cells in at least \code{min_samples}
samples in at least one condition.}

\item{min_samples}{Filtering parameter. Default = \code{min(table(group_IDs)) - 1},
i.e. one less than the number of samples in the smallest group. Clusters are kept for
differential testing if they have at least \code{min_cells} cells in at least
\code{min_samples} samples in at least one condition.}
}
\value{
Returns a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} object,
  with differential test results stored in the \code{rowData} slot. Results include raw
  p-values and adjusted p-values, which can be used to rank clusters by evidence for
  differential abundance. The results can be accessed with the
  \code{\link[SummarizedExperiment]{rowData}} accessor function.
}
\description{
Calculate tests for differential abundance of clusters using method 'diffcyt-DA-GLMM'
}
\details{
Calculates tests for differential abundance of clusters, using generalized linear mixed
models (GLMMs) for each cluster. This methodology was originally developed and
described by Nowicka et al. (2017), \emph{F1000Research}.

For more details on the underlying statistical methodology, refer to the paper by
Nowicka et al. (2017), \emph{F1000Research}. The implementation here contains several
additional modifications. In particular, we use high-resolution clustering and report
results at the high-resolution cluster level, instead of relying on a manual
cluster-merging step. We also include a filtering step to remove high-resolution
clusters with very small numbers of cells, to improve power.

The experimental design must be specified using a model formula, which can be created
with \code{\link{createFormula}}. Flexible experimental designs are possible, including
batch effects, continuous covariates, and blocking (e.g. for paired designs). Random
intercept terms are included for blocks (e.g. paired designs), as well as
'observation-level random effects' to account for overdispersion typically seen in
high-dimensional cytometry data (Nowicka et al., 2017, \emph{F1000Research}). See
\code{\link{createFormula}} for more details.

The contrast matrix specifying the contrast of interest can be created with
\code{\link{createContrast}}. See \code{\link{createContrast}} for more details.

Filtering: Clusters are kept for differential testing if they have at least
\code{min_cells} cells in at least \code{min_samples} samples in at least one
condition. This removes clusters with very low cell counts across conditions, which
improves power.
}
\examples{
# to do

}
\seealso{
to do
}
