% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDS_LMM.R
\name{testDS_LMM}
\alias{testDS_LMM}
\title{Test for differential functional states: method 'diffcyt-DS-LMM'}
\usage{
testDS_LMM(d_counts, d_medians, formula, contrast, min_cells = 3,
  min_samples = NULL)
}
\arguments{
\item{d_counts}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster cell counts, from \code{\link{calcCounts}}.}

\item{d_medians}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing cluster medians (median expression of each marker for each cluster-sample
combination), from \code{\link{calcMedians}}.}

\item{formula}{Model formula object, created with \code{\link{createFormula}}. This
should be a list containing three elements: \code{formula}, \code{data}, and
\code{random_terms}: the model formula, data frame of corresponding variables, and
variable indicating whether the model formula contains random effect terms. See
\code{\link{createFormula}} for details.}

\item{contrast}{Contrast matrix, created with \code{\link{createContrast}}. See
\code{\link{createContrast}} for details.}

\item{min_cells}{Filtering parameter. Default = 3. Clusters are kept for differential
testing if they have at least \code{min_cells} cells in at least \code{min_samples}
samples in at least one condition.}

\item{min_samples}{Filtering parameter. Default = \code{min(table(group_IDs)) - 1},
i.e. one less than the number of samples in the smallest group. Clusters are kept for
differential testing if they have at least \code{min_cells} cells in at least
\code{min_samples} samples in at least one condition.}
}
\value{
Returns a new \code{\link[SummarizedExperiment]{SummarizedExperiment}} object,
  where rows = cluster-marker combinations, and columns = samples. In the rows,
  clusters are repeated for each functional state marker (i.e. the sheets or 'assays'
  from the previous \code{d_medians} object are stacked into a single matrix).
  Differential test results are stored in the \code{rowData} slot. Results include raw
  p-values and adjusted p-values, which can be used to rank cluster-marker combinations
  by evidence for differential functional states. The results can be accessed with the
  \code{\link[SummarizedExperiment]{rowData}} accessor function.
}
\description{
Calculate tests for differential functional states of clusters using method
'diffcyt-DS-LMM'
}
\details{
Calculates tests for differential functional states of clusters (i.e. differential
expression of functional state markers within clusters), using linear mixed models
(LMMs). Clusters are defined using cell type markers, and functional states are
characterized by the median transformed expression of functional state markers.

This methodology was originally developed and described by Nowicka et al. (2017),
\emph{F1000Research}, and has been modified here to make use of high-resolution
clustering to enable investigation of rare cell populations. Note that unlike the
original method by Nowicka et al., we do not attempt to manually merge clusters into
canonical cell populations. Instead, results are reported at the high-resolution
cluster level, and the interpretation of significant differential clusters is left to
the user via visualizations such as heatmaps and tSNE plots (see the package vignette
for a detailed example).

This method fits linear mixed models (LMMs) for each cluster-marker combination
(functional state markers only), and calculates differential tests separately for each
cluster-marker combination. The response variable in each model is the median
arcsinh-transformed marker expression of the functional state marker, which is assumed
to follow a Gaussian distribution. There is one model per cluster per functional state
marker. Within each model, sample-level weights are included for the number of cells
per sample; these weights represent the relative uncertainty in calculating each median
value. (Additional uncertainty exists due to variation in the total number of cells per
cluster; however, it is not possible to account for this, since there are separate
models for each cluster-marker combination. By contrast, the 'diffcyt-DS-limma' method
can account for this source of uncertainty using empirical Bayes methodology.) We also
include a filtering step to remove clusters with very small numbers of cells, to
improve statistical power.

For more details on the statistical methodology, see Nowicka et al. (2017),
\emph{F1000Research} (section 'Differential analysis of marker expression stratified by
cell population'.)

The experimental design must be specified using a model formula, which can be created
with \code{\link{createFormula}}. Flexible experimental designs are possible, including
blocking (e.g. paired designs), batch effects, and continuous covariates. Blocking
variables can be included as either random intercept terms or fixed effect terms (see
\code{\link{createFormula}}). For paired designs, we recommend using random intercept
terms to improve statistical power; see Nowicka et al. (2017), \emph{F1000Research} for
details. Batch effects and continuous covariates should be included as fixed effects.

If no random intercept terms are included in the model formula, model fitting is
performed using a linear model (LM) instead of a LMM.

The contrast matrix specifying the contrast of interest can be created with
\code{\link{createContrast}}. See \code{\link{createContrast}} for more details.

Filtering: Cluster-marker combinations are kept for differential testing if they have
at least \code{min_cells} cells in at least \code{min_samples} samples in at least one
condition. This removes clusters with very low cell counts across conditions, which
improves power.
}
\examples{
# A full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline on an experimental data set is available in the package vignette.

}
