% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_transformData.R
\name{transformData}
\alias{transformData}
\title{Transform data}
\usage{
transformData(d_se, cofactor = 5)
}
\arguments{
\item{d_se}{Input data. Assumed to be in the form of a
\code{\link[SummarizedExperiment]{SummarizedExperiment}}, prepared with the function
\code{\link{prepareData}}. Column meta-data is assumed to contain a vector of logical
entries (\code{is_marker}) indicating protein marker columns.}

\item{cofactor}{Cofactor parameter for 'arcsinh' transform. Default = 5, which is
appropriate for mass cytometry (CyTOF) data. For fluorescence flow cytometry, we
recommend cofactor = 150 instead.}
}
\value{
\code{d_se}: Data with transform applied to protein marker columns.
}
\description{
Transform data prior to clustering
}
\details{
Flow and mass cytometry data should be transformed prior to clustering. The raw data
follows an approximately log-normal distribution. Transforming with a log (or similar)
function brings the data closer to a normal distribution, which improves clustering
performance and allows positive and negative populations to be distinguished more
clearly.

This function implements an inverse hyperbolic sine ('arcsinh') transform with
adjustable 'cofactor' parameter. The arcsinh transform is widely used for CyTOF data.
It behaves similarly to a log transform at high values, but is approximately linear
near zero; so unlike the log, it can handle zeros or small negative values. The
cofactor parameter controls the width of the linear region. Zero values and small
negatives occur in CyTOF data when no ions are detected in a given channel (negatives
are due to background subtraction and randomization of integer count values, which are
performed by default by the CyTOF instrument software).

Recommended values for the cofactor parameter are 5 (mass cytometry, CyTOF) or 150
(fluorescence flow cytometry); see Bendall et al. (2011), \emph{Science}, Supplementary
Figure S2.

The transform should be applied to protein marker columns only. The
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object created in the previous
step (\code{\link{prepareData}}) is assumed to contain a vector of logical entries
named \code{is_marker} in the column meta-data, indicating whether each column is a
protein marker. (If this is not available, all columns will be transformed instead.)
}
\examples{
# For a full workflow example demonstrating the use of each function in the 'diffcyt'
# pipeline, see the package vignette.

# Create some random data (without differential signal)
cofactor <- 5
set.seed(123)
d_input <- list(
  sample1 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample2 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample3 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor, 
  sample4 = sinh(matrix(rnorm(20000, mean = 0, sd = 1), ncol = 20)) * cofactor
)

sample_info <- data.frame(
  sample = factor(paste0("sample", 1:4)), 
  group = factor(c("group1", "group1", "group2", "group2")), 
  stringsAsFactors = FALSE
)

marker_info <- data.frame(
  marker_name = paste0("marker", 1:20), 
  is_marker = rep(TRUE, 20), 
  is_type_marker = c(rep(TRUE, 10), rep(FALSE, 10)), 
  is_state_marker = c(rep(FALSE, 10), rep(TRUE, 10)), 
  stringsAsFactors = FALSE
)

# Prepare data
d_se <- prepareData(d_input, sample_info, marker_info)

# Transform data
d_se <- transformData(d_se)

}
